fix(review): effective_code_size returns u64 to avoid truncation
eMMC chips can be larger than 4 GiB (e.g., 8 GiB = 0x200000000).
The trait method returned u32, which would truncate the detected
capacity. Changed to u64. The AtomicU64 storage was already correct;
only the trait return type and T76 override needed updating.

+++ crates/minipro-core/src/protocol/mod.rs
@@ -60,8 +60,8 @@ pub trait Protocol: Send + Sync {
     /// Defaults to `device.code_memory_size`.  T76Protocol overrides this
     /// for eMMC, where the database value is a placeholder (0x200) and the
     /// real capacity is detected at runtime from EXT_CSD.
-    fn effective_code_size(&self, device: &Device) -> u32 {
-        device.code_memory_size
+    fn effective_code_size(&self, device: &Device) -> u64 {
+        device.code_memory_size as u64
     }
 
     /// Read one block of memory into `ds.data`.
